src/requirements.txt
azure.core
azure.ai.documentintelligence
streamlit
azure.storage.blob
python-dotenv